Update account wallet saving and loading from the file
