+++ tools/evil-spammer/accountwallet/config.go
@@ -3,6 +3,8 @@ package accountwallet
 import (
 	"os"
 
+	"github.com/google/martian/log"
+
 	"github.com/iotaledger/hive.go/ds/types"
 	"github.com/iotaledger/hive.go/ierrors"
 	iotago "github.com/iotaledger/iota.go/v4"
@@ -47,7 +49,7 @@ type AllotAccountParams struct {
 // TODO do wee need to restrict that only one instance of the wallet runs?
 const lockFile = "wallet.LOCK"
 
-type AccountData struct {
+type StateData struct {
 	Accounts []Account `serix:"0,mapKey=accounts,lengthPrefixType=uint8"`
 	// TODO: other info that the account wallet needs to store
 }
@@ -58,38 +60,17 @@ type Account struct {
 	// TODO: other info of an account
 }
 
-func loadWallet(filename string) (*AccountWallet, error) {
-	wallet := NewAccountWallet()
-
-	//walletStateBytes, err := os.ReadFile(filename)
-	//if err != nil {
-	//	return nil, ierrors.Wrap(err, "failed to read wallet file")
-	//}
-	//
-	//var data AccountData
-	//_, err = wallet.api.Decode(bytes, &data)
-	//if err != nil {
-	//	return nil, ierrors.Wrap(err, "failed to decode from file")
-	//}
-	//
-	//TODO: import the data to Wallet
-	//wallet.FromAccountData(data)
+func loadWallet(filename string, api iotago.API) (wallet *AccountWallet, err error) {
+	walletStateBytes, err := os.ReadFile(filename)
+	if err != nil {
+		log.Infof("No working wallet file %s found, creating new wallet...", filename)
+		wallet = NewAccountWallet(api)
+	} else {
+		wallet, err = AccountWalletFromBytes(api, walletStateBytes)
+		if err != nil {
+			return nil, ierrors.Wrap(err, "failed to create wallet from bytes")
+		}
+	}
 
 	return wallet, nil
-}
-
-func writeWalletStateFile(w *AccountWallet, filename string) error {
-	dataToDump := w.ToAccountData()
-
-	bytesToWrite, err := w.api.Encode(dataToDump)
-	if err != nil {
-		return ierrors.Wrap(err, "failed to encode account data")
-	}
-
-	//nolint:gosec // users should be able to read the file
-	if err = os.WriteFile(filename, bytesToWrite, 0o644); err != nil {
-		return ierrors.Wrap(err, "failed to write account file")
-	}
-
-	return nil
 }
